+++ android/companion/src/com/bunnytasker/R.java
@@ -17,6 +17,7 @@ public final class R {
     public static final int crown_gold=0x7f020001;
     public static final int ic_launcher=0x7f020002;
     public static final int ic_launcher_foreground=0x7f020003;
+    public static final int ic_launcher_monochrome=0x7f020004;
   }
   public static final class id {
     public static final int app_icon=0x7f030000;
@@ -95,7 +96,6 @@ public final class R {
   }
   public static final class mipmap {
     public static final int ic_launcher=0x7f050000;
-    public static final int ic_launcher_foreground=0x7f050001;
   }
   public static final class string {
     public static final int app_name=0x7f060000;